Fix VC16 warnings

+++ Encoder/src/Encoder_prepareSequence.cpp
@@ -63,7 +63,7 @@ auto Encoder::prepareSequence(IvSequenceParams sourceIvs)
   setGiGeometry3dCoordinatesBitdepthMinus1();
 
   // Update views per atlas info
-  m_ivs.mvpl().mvp_num_views_minus1(m_isBasicView.size() - 1);
+  m_ivs.mvpl().mvp_num_views_minus1(uint16_t(m_isBasicView.size() - 1));
   m_ivs.mvpl().mvp_atlas_count_minus1(m_ivs.vps.vps_atlas_count_minus1());
   for (uint8_t a = 0; a <= m_ivs.vps.vps_atlas_count_minus1(); ++a)
     for (uint16_t v = 0; v <= m_ivs.mvpl().mvp_num_views_minus1(); ++v) {
